+++ src/zakuro.cpp
@@ -293,6 +293,8 @@ int Zakuro::Run() {
     sorac_config.client_key = config_.client_key;
     sorac_config.signaling_urls = config_.sora_signaling_urls;
     sorac_config.channel_id = config_.sora_channel_id;
+    sorac_config.disable_signaling_url_randomization =
+        config_.sora_disable_signaling_url_randomization;
     sorac_config.video = config_.sora_video;
     sorac_config.audio = config_.sora_audio;
     sorac_config.video_codec_type = config_.sora_video_codec_type;